# orc8r/cloud/go/services/orchestrator/obsidian/models/swagger.v1.yml
...
swagger: '2.0'
...
info:
  title: Orchestrator Network Management
  description: Orchestrator REST APIs
  version: 1.0.0
...
basePath: /magma/v1
paths:
...
  /networks/{network_id}/gateways/{gateway_id}/command/reboot:
    post:
      summary: Reboot gateway device
      tags:
        - Commands
      parameters:
        - $ref: './orc8r-swagger-common.yml#/parameters/network_id'
        - $ref: './orc8r-swagger-common.yml#/parameters/gateway_id'
      responses:
        '200':
          description: Success
        default:
          $ref: './orc8r-swagger-common.yml#/responses/UnexpectedError'
  /networks/{network_id}/gateways/{gateway_id}/command/restart_services:
    post:
      summary: Restart gateway services
      tags:
        - Commands
      parameters:
        - $ref: './orc8r-swagger-common.yml#/parameters/network_id'
        - $ref: './orc8r-swagger-common.yml#/parameters/gateway_id'
        - in: body
          name: Services
          description: Services
          required: true
          schema:
            type: array
            items:
              type: string
            example: []
      responses:
        '200':
          description: Success
        default:
          $ref: './orc8r-swagger-common.yml#/responses/UnexpectedError'
...
  /networks/{network_id}/gateways/{gateway_id}/command/generic:
    post:
      summary: Execute generic command on gateway
      tags:
        - Commands
      parameters:
        - $ref: './orc8r-swagger-common.yml#/parameters/network_id'
        - $ref: './orc8r-swagger-common.yml#/parameters/gateway_id'
        - in: body
          name: Parameters
          description: Parameters
          required: true
          schema:
            $ref: '#/definitions/generic_command_params'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/generic_command_response'
        default:
          $ref: './orc8r-swagger-common.yml#/responses/UnexpectedError'
...
definitions:
...
  generic_command_params:
    type: object
    required:
      - command
    properties:
      command:
        type: string
        minLength: 1
        example: bash
      params:
        type: object
        additionalProperties:
          type: object
        example: {"shell_params": ["-c 'ls -1 .'"]}
  generic_command_response:
    type: object
    properties:
      response:
        type: object
        additionalProperties:
          type: object
        example: {}
...